Prefer loading site_init templates from flat subdirectories. As a fallback, load a zipfile.
In development, we don't want to have to be loading the site_init templates from zip files as they will probably be stale. We also don't want to always regenrate them as that seems excessive. In production, the flat file versions won't exist, so we'll fall back to loading the zip file.