Use an incrementing number instead of the Wordpress ID. Get rid of funky characters. Change spaces into underscores.
By default the posts generated from the database get the Wordpress ID in
the form of "NNNN. <post title>.html." While the ID is useful to WP it
isn't useful to the blogger using Blogofile. This change instead assigns
the number in the filename with an incrementing integer (with a scale of
4).

The post title itself requires some special attention as well.
Special characters such as ?/-:'!* can also be present in titles,
but these can be troublesome when on a filesystem. This change removes
them. Spaces are converted to underscores (_) to retain a somewhat
readable name.

The end result is each WP post being written in chronological order
(your first post is 0001_<post title>.html) and with readable
file names.