Makes the blog controller templates relocatable.
Blog templates from 'simple_blog init' are now placed in the _templates/blog directory to differentiate them from site templates. This was very confusing for new users previously. Hopefully the new structure has more logical sense to it.

This redefines the parent feature of the templates, some are children of simple_blog, some are children of blog_features:

Children of simple_blog (goes in _templates):

 * base.mako
 * site.mako
 * head.mako
 * header.mako
 * footer.mako

Children of blog_features (goes in _templates/blog):

 * atom.mako
 * chronological.mako
 * permapage.mako
 * post_excerpt.mako
 * post.mako
 * rss.mako

blog_feature templates will still work from the _templates directory if they are not found in the _templates/blog directory. This is for two reasons:

 1) All existing blogofile users will have these templates in the _templates dir.
 2) The _templates dir has to be on the lookup path in order for inheritance to work anyway for the base template(s).

Finally, there are new blog config options:

The blog template directory can be relocated anywhere in your source directory:

    controllers.blog.template_path = "_my_templates/my_blog_templates"

The base template for the blog templates no longer has to be named site.mako:

    controller.blog.base_template = "some_other_base.mako"